+++ src/main/java/com/epam/reportportal/listeners/ListenerParameters.java
@@ -39,8 +39,6 @@ public class ListenerParameters implements Cloneable {
 
 	private static final int DEFAULT_REPORTING_TIMEOUT = 5 * 60;
 	private static final int DEFAULT_IO_POOL_SIZE = 100;
-	private static final int DEFAULT_MAX_CONNECTIONS_PER_ROUTE = 50;
-	private static final int DEFAULT_MAX_CONNECTIONS_TOTAL = 100;
 	private static final boolean DEFAULT_ENABLE = true;
 	private static final boolean DEFAULT_SKIP_ISSUE = true;
 	private static final boolean DEFAULT_CONVERT_IMAGE = false;
@@ -52,6 +50,7 @@ public class ListenerParameters implements Cloneable {
 	private static final String DEFAULT_LOCK_FILE_NAME = "reportportal.lock";
 	private static final String DEFAULT_SYNC_FILE_NAME = "reportportal.sync";
 	private static final long DEFAULT_FILE_WAIT_TIMEOUT_MS = TimeUnit.MINUTES.toMillis(1);
+	private static final int DEFAULT_RX_BUFFER_SIZE = 128;
 
 	private String description;
 	private String apiKey;
@@ -80,11 +79,13 @@ public class ListenerParameters implements Cloneable {
 	private String syncFileName;
 	private long fileWaitTimeout;
 
+	private int rxBufferSize;
+
 	public ListenerParameters() {
 
 		this.isSkippedAnIssue = DEFAULT_SKIP_ISSUE;
 
-		this.batchLogsSize = LoggingContext.DEFAULT_BUFFER_SIZE;
+		this.batchLogsSize = LoggingContext.DEFAULT_LOG_BATCH_SIZE;
 		this.convertImage = DEFAULT_CONVERT_IMAGE;
 		this.reportingTimeout = DEFAULT_REPORTING_TIMEOUT;
 		this.httpLogging = DEFAULT_HTTP_LOGGING;
@@ -102,6 +103,7 @@ public class ListenerParameters implements Cloneable {
 		this.lockFileName = DEFAULT_LOCK_FILE_NAME;
 		this.syncFileName = DEFAULT_SYNC_FILE_NAME;
 		this.fileWaitTimeout = DEFAULT_FILE_WAIT_TIMEOUT_MS;
+		this.rxBufferSize = DEFAULT_RX_BUFFER_SIZE;
 	}
 
 	public ListenerParameters(PropertiesLoader properties) {
@@ -116,7 +118,7 @@ public class ListenerParameters implements Cloneable {
 		this.enable = properties.getPropertyAsBoolean(ENABLE, DEFAULT_ENABLE);
 		this.isSkippedAnIssue = properties.getPropertyAsBoolean(SKIPPED_AS_ISSUE, DEFAULT_SKIP_ISSUE);
 
-		this.batchLogsSize = properties.getPropertyAsInt(BATCH_SIZE_LOGS, LoggingContext.DEFAULT_BUFFER_SIZE);
+		this.batchLogsSize = properties.getPropertyAsInt(BATCH_SIZE_LOGS, LoggingContext.DEFAULT_LOG_BATCH_SIZE);
 		this.convertImage = properties.getPropertyAsBoolean(IS_CONVERT_IMAGE, DEFAULT_CONVERT_IMAGE);
 		this.reportingTimeout = properties.getPropertyAsInt(REPORTING_TIMEOUT, DEFAULT_REPORTING_TIMEOUT);
 		this.httpLogging = properties.getPropertyAsBoolean(HTTP_LOGGING, DEFAULT_HTTP_LOGGING);
@@ -135,6 +137,8 @@ public class ListenerParameters implements Cloneable {
 		this.lockFileName = properties.getProperty(LOCK_FILE_NAME, DEFAULT_LOCK_FILE_NAME);
 		this.syncFileName = properties.getProperty(SYNC_FILE_NAME, DEFAULT_SYNC_FILE_NAME);
 		this.fileWaitTimeout = properties.getPropertyAsInt(FILE_WAIT_TIMEOUT_MS, (int) DEFAULT_FILE_WAIT_TIMEOUT_MS);
+
+		this.rxBufferSize = properties.getPropertyAsInt(RX_BUFFER_SIZE, DEFAULT_RX_BUFFER_SIZE);
 	}
 
 	public String getDescription() {
@@ -337,6 +341,17 @@ public class ListenerParameters implements Cloneable {
 		this.httpLogging = httpLogging;
 	}
 
+	public int getRxBufferSize() {
+		return ofNullable(System.getProperty("rx2.buffer-size"))
+				.map(Integer::valueOf)
+				.map(s->Math.max(1, s))
+				.orElse(rxBufferSize);
+	}
+
+	public void setRxBufferSize(int size) {
+		rxBufferSize = size;
+	}
+
 	@VisibleForTesting
 	Mode parseLaunchMode(String mode) {
 		return Mode.isExists(mode) ? Mode.valueOf(mode.toUpperCase()) : Mode.DEFAULT;
@@ -367,6 +382,7 @@ public class ListenerParameters implements Cloneable {
 
 	@Override
 	public String toString() {
+		@SuppressWarnings("StringBufferReplaceableByString")
 		final StringBuilder sb = new StringBuilder("ListenerParameters{");
 		sb.append("description='").append(description).append('\'');
 		sb.append(", apiKey='").append(apiKey).append('\'');
@@ -393,6 +409,7 @@ public class ListenerParameters implements Cloneable {
 		sb.append(", lockFileName=").append(lockFileName);
 		sb.append(", syncFileName=").append(syncFileName);
 		sb.append(", fileWaitTimeout=").append(fileWaitTimeout);
+		sb.append(", rxBufferSize=").append(rxBufferSize);
 		sb.append('}');
 		return sb.toString();
 	}