+++ src/main/java/com/epam/reportportal/service/LaunchImpl.java
@@ -190,8 +190,7 @@ public class LaunchImpl extends Launch {
 		LaunchLoggingContext.init(this.launch,
 				this.rpClient,
 				getScheduler(),
-				getParameters().getBatchLogsSize(),
-				getParameters().isConvertImage()
+				getParameters()
 		);
 		getAnalyticsService().sendEvent(launch, startRq);
 		return this.launch;
@@ -249,7 +248,7 @@ public class LaunchImpl extends Launch {
 
 		item.subscribeOn(getScheduler()).subscribe(logMaybeResults("Start test item"));
 		QUEUE.getUnchecked(item).addToQueue(item.ignoreElement().onErrorComplete());
-		LoggingContext.init(launch, item, rpClient, getScheduler(), getParameters().getBatchLogsSize(), getParameters().isConvertImage());
+		LoggingContext.init(launch, item, rpClient, getScheduler(), getParameters());
 
 		getStepReporter().setParent(item);
 		StepAspect.setParentId(item);
@@ -281,7 +280,7 @@ public class LaunchImpl extends Launch {
 		})).cache();
 		item.subscribeOn(getScheduler()).subscribe(logMaybeResults("Start test item"));
 		QUEUE.getUnchecked(item).withParent(parentId).addToQueue(item.ignoreElement().onErrorComplete());
-		LoggingContext.init(launch, item, rpClient, getScheduler(), getParameters().getBatchLogsSize(), getParameters().isConvertImage());
+		LoggingContext.init(launch, item, rpClient, getScheduler(), getParameters());
 
 		getStepReporter().setParent(item);
 		StepAspect.setParentId(item);